Eliminar horario
Ya se pueden eliminar horarios

+++ publicHTML/backend/perfil/subirIMG.php
@@ -1,22 +1,27 @@
 <?php 
-session_start();
+
 include('../conexion.php');
 
-if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-    if (isset($_SESSION['paciente']) && !isset($_SESSION['odontologo'])) {
-        subirPacienteIMG($pdo);
-    } else if (isset($_SESSION['odontologo']) && !isset($_SESSION['paciente'])) {
-        subirOdontologoIMG($pdo);
-    } else {
-        echo "Fuera";
-    }
-    exit();
-} else {
-    exit();
-}
+session_start();
 
-function checarIMG($pdo) {
+if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+
+    if (isset($_SESSION['paciente']) && !isset($_SESSION['odontologo'])) subirPacienteIMG();
+    
+    else if (isset($_SESSION['odontologo']) && !isset($_SESSION['paciente'])) subirOdontologoIMG();
+    
+    else echo "Fuera";
+    
+    exit();
+} 
+else exit();
+
+function checarIMG() {
+
+    global $pdo;
+
     $estado = array(
+
         'ok' => false,
         'nombre' => ''
     );
@@ -31,7 +36,8 @@ function checarIMG($pdo) {
             $estado['nombre'] = $stmt->fetch(PDO::FETCH_ASSOC)['foto'];
             $estado['ok'] = true;
         }
-    } else if (isset($_SESSION['odontologo']) && !isset($_SESSION['paciente'])) {
+    } 
+    else if (isset($_SESSION['odontologo']) && !isset($_SESSION['paciente'])) {
 
         $consulta = "SELECT foto FROM odontologo WHERE idodontologo = :ido";
         $stmt = $pdo->prepare($consulta);
@@ -41,19 +47,21 @@ function checarIMG($pdo) {
             $estado['nombre'] = $stmt->fetch(PDO::FETCH_ASSOC)['foto'];
             $estado['ok'] = true;
         }
-    } else {
-        return false;
     }
+    else return false;
 
     return $estado;
 }
 
-function subirPacienteIMG($pdo) {
+function subirPacienteIMG() {
+
+    global $pdo;
+
     $respuesta = array();
 
     // Obtengo la imagen del ajax
     $file = $_FILES['file'];
-    $ruta_carpeta = $_SERVER['DOCUMENT_ROOT'] . "/Proyecto-3-BF/publicHTML/img/imgPerfil/";
+    $ruta_carpeta = $_SERVER['DOCUMENT_ROOT'] . "/Proyecto-3-BF/publicHTML/backend/almacenamiento/fotosdeperfil/";
     
     // Genera un nombre único para que la imagen no se encuentre repetida
     $extension = pathinfo($file['name'], PATHINFO_EXTENSION);
@@ -62,14 +70,16 @@ function subirPacienteIMG($pdo) {
 
     // Intenta guardar la imagen
     try {
+
         if (move_uploaded_file($file['tmp_name'], $ruta_guardar_archivo)) {
+
             $respuesta['enviar'] = "Se cambió su perfil";
+
             try {
+
                 //Borro la imagen vieja del usuario
-                $estadoIMG = checarIMG($pdo);
-                if ($estadoIMG['ok'] && file_exists($ruta_carpeta . $estadoIMG['nombre'])) {
-                    unlink($ruta_carpeta . $estadoIMG['nombre']);
-                }
+                $estadoIMG = checarIMG();
+                if ($estadoIMG['ok'] && file_exists($ruta_carpeta . $estadoIMG['nombre'])) unlink($ruta_carpeta . $estadoIMG['nombre']);
 
                 // Consulta para modificar paciente
                 $consulta = "UPDATE paciente SET foto = :img WHERE idpaciente = :idp";
@@ -78,29 +88,30 @@ function subirPacienteIMG($pdo) {
                 $stmt->bindParam(':idp', $_SESSION['paciente']['idpaciente']);
                 $stmt->execute();
                 reloadSession();
-            } catch (PDOException $e) {
+            } 
+            catch (PDOException $e) {
+
                 $respuesta['error'] = "Ha ocurrido un error: " . $e->getMessage();
             }
-        } else {
-            $respuesta['error'] = "Error al mover el archivo. Verifica los permisos de la carpeta.";
-        }
-    } catch (Exception $e) {
+        } else $respuesta['error'] = "Error al mover el archivo. Verifica los permisos de la carpeta.";
+    } 
+    catch (Exception $e) {
+
         $respuesta['error'] = "Excepción: " . $e->getMessage();
     }
-
     header('Content-Type: application/json');
     echo json_encode($respuesta);
 }
 
+function subirOdontologoIMG() {
 
+    global $pdo;
 
-
-function subirOdontologoIMG($pdo){
     $respuesta = array();
 
     // Obtengo la imagen del ajax
     $file = $_FILES['file'];
-    $ruta_carpeta = $_SERVER['DOCUMENT_ROOT'] . "/Proyecto-3-BF/publicHTML/img/imgPerfil/";
+    $ruta_carpeta = $_SERVER['DOCUMENT_ROOT'] . "/Proyecto-3-BF/publicHTML/backend/almacenamiento/fotosdeperfil/";
     
     // Genera un nombre único para que la imagen no se encuentre repetida
     $extension = pathinfo($file['name'], PATHINFO_EXTENSION);
@@ -109,14 +120,16 @@ function subirOdontologoIMG($pdo){
 
     // Intenta guardar la imagen
     try {
+
         if (move_uploaded_file($file['tmp_name'], $ruta_guardar_archivo)) {
+
             $respuesta['enviar'] = "Se cambió su perfil";
+
             try {
+                
                 //Borro la imagen vieja del usuario
-                $estadoIMG = checarIMG($pdo);
-                if ($estadoIMG['ok'] && file_exists($ruta_carpeta . $estadoIMG['nombre'])) {
-                    unlink($ruta_carpeta . $estadoIMG['nombre']);
-                }
+                $estadoIMG = checarIMG();
+                if ($estadoIMG['ok'] && file_exists($ruta_carpeta . $estadoIMG['nombre'])) unlink($ruta_carpeta . $estadoIMG['nombre']);
 
                 // Consulta para modificar paciente
                 $consulta = "UPDATE odontologo SET foto = :img WHERE idodontologo = :ido";
@@ -127,18 +140,18 @@ function subirOdontologoIMG($pdo){
                 reloadSession();
 
             } catch (PDOException $e) {
+
                 $respuesta['error'] = "Ha ocurrido un error: " . $e->getMessage();
             }
-        } else {
-            $respuesta['error'] = "Error al mover el archivo. Verifica los permisos de la carpeta.";
-        }
-    } catch (Exception $e) {
+        } 
+        else $respuesta['error'] = "Error al mover el archivo. Verifica los permisos de la carpeta.";
+    } 
+    catch (Exception $e) {
+
         $respuesta['error'] = "Excepción: " . $e->getMessage();
     }
-
     header('Content-Type: application/json');
     echo json_encode($respuesta);
 }
-
 
 ?>
